правки в модуль scale
